Fix file upload
The `<form ng-submit…>` was being triggered when the button for
selecting files was pressed (prior to any file selection). This in turn
caused the `.inprocess` flag to be set to `true`, deactivating the
submit button.

Removing the `<form ng-submit…>` and making use of `<button …
ng-click=“uploadFliles()”>` on the submit button fixes this.

This addresses
https://github.com/joni2back/angular-filemanager/issues/204